refactored printing component display
