+++ src/Pages/2DPrinting/2DPrintingHealthCheck.js
@@ -1,28 +1,11 @@
 import React, { useState, useEffect } from 'react';
-import { healthCheck } from '../../APIFunctions/2DPrinting';
 import {
   Spinner,
   Container
 } from 'reactstrap';
 
 export default function PrintingHealthCheck(props) {
-  const [loading, setLoading] = useState(true);
-  const [printerHealthy, setPrinterHealthy] = useState(false);
-
-  async function checkPrinterHealth() {
-    setLoading(true);
-    const status = await healthCheck();
-    if (status && !status.error) {
-      setPrinterHealthy(true);
-    } else {
-      setPrinterHealthy(false);
-    }
-    setLoading(false);
-  }
-
-  useEffect(() => {
-    checkPrinterHealth();
-  }, []);
+  const {printerHealthy, loading} = props;
 
   function renderPrinterHealth() {
     if (loading) {
